# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "xg backend rest api",
        "title": "xg REST API",
        "termsOfService": "https://localhost:8088/v1",
        "contact": {},
        "version": "0.1 alpha"
    },
    "paths": {
...
        "/api/org": {
            "post": {
                "description": "create org",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "createOrg",
                "parameters": [
                    {
                        "description": "create org request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateOrgWithSubOrgsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/org/{id}": {
            "get": {
                "description": "get org by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "getOrgById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Org"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/org/{id}/review/approve": {
            "put": {
                "description": "approve org",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "approveOrg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/org/{id}/review/reject": {
            "put": {
                "description": "reject org",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "rejectOrg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/org/{id}/revoke": {
            "put": {
                "description": "revoke org",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "revokeOrg",
                "parameters": [
                    {
                        "type": "string",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/org/{id}/subjects": {
            "get": {
                "description": "get org subjects by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "getOrgSubjectsById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.SubjectsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/orgs": {
            "get": {
                "description": "list all organizations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "listOrgs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.OrgListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/orgs/campus": {
            "get": {
                "description": "search sub organizations with condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "searchSubOrgs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search organizations with subjects",
                        "name": "subjects",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search organizations with address",
                        "name": "address",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.OrgListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/orgs/pending": {
            "get": {
                "description": "list pending organizations to check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "listPendingOrgs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.OrgListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateOrgRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "entity.CreateOrgWithSubOrgsRequest": {
            "type": "object",
            "properties": {
                "org": {
                    "$ref": "#/definitions/entity.CreateOrgRequest"
                },
                "sub_orgs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CreateOrgRequest"
                    }
                }
            }
        },
...
        "entity.Org": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "sub_orgs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Org"
                    }
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "route.IdResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "route.OrgListResponse": {
            "type": "object",
            "properties": {
                "orgs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Org"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "route.Response": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                }
            }
        },
        "route.SubjectsResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}